+++ cmr/__init__.py
@@ -1,3 +1,25 @@
-from .queries import GranuleQuery, CollectionQuery, ToolQuery, ServiceQuery, VariableQuery, CMR_OPS, CMR_UAT, CMR_SIT
+from .queries import (
+    CMR_OPS,
+    CMR_SIT,
+    CMR_UAT,
+    CollectionQuery,
+    DayNightFlag,
+    GranuleQuery,
+    Query,
+    ServiceQuery,
+    ToolQuery,
+    VariableQuery,
+)
 
-__all__ = ["GranuleQuery", "CollectionQuery", "ToolQuery", "ServiceQuery", "VariableQuery", "CMR_OPS", "CMR_UAT", "CMR_SIT"]
+__all__ = [
+    "CMR_OPS",
+    "CMR_SIT",
+    "CMR_UAT",
+    "CollectionQuery",
+    "DayNightFlag",
+    "GranuleQuery",
+    "Query",
+    "ServiceQuery",
+    "ToolQuery",
+    "VariableQuery",
+]