+++ CHANGELOG.md
@@ -8,6 +8,22 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.12.0]
+
+### Added
+
+- Support searching for cloud-hosted collections
+  ([#54](https://github.com/nasa/python_cmr/issues/54))
+- Option to search for collections and granules by `revision_date` ([#67](https://github.com/nasa/python_cmr/issues/67))
+
+### Fixed
+
+- Do not clobber other headers when a query's `Authorization` header is set via
+  the `bearer_token` or `token` method.  However, if multiple calls are made to
+  the `bearer_token` and/or `token` methods, later calls overwrite the
+  `Authorization` setting of prior calls (i.e., the last call wins).
+  ([#48](https://github.com/nasa/python_cmr/issues/48))
+- Export `DayNightFlag` type alias ([#51](https://github.com/nasa/python_cmr/issues/51))
 
 ## [0.11.0]
 
@@ -116,7 +132,9 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Prior releases of this software originated from
   <https://github.com/jddeal/python-cmr/releases>
 
-[Unreleased]: https://github.com/nasa/python_cmr/compare/v0.10.0...HEAD
+[Unreleased]: https://github.com/nasa/python_cmr/compare/v0.12.0...HEAD
+[0.12.0]: https://github.com/nasa/python_cmr/compare/v0.11.0...v0.12.0
+[0.11.0]: https://github.com/nasa/python_cmr/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/nasa/python_cmr/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/nasa/python_cmr/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/nasa/python_cmr/compare/v0.7.0...v0.8.0